refactor: isolate ready chunk body helper
