+++ actix-http/src/test/mod.rs
@@ -1,5 +1,7 @@
 //! Various testing helpers for use in internal and app tests.
 
+#[cfg(test)]
+mod ready_chunk_body;
 mod test_buffer;
 mod test_request;
 mod test_seq_buffer;
